+++ source/ambient.d.ts
@@ -5,3 +5,4 @@ declare module "json2ts"
 
 declare module "async-exit-hook"
 declare module "graphql-resolvers"
+declare module "graphql-tools-types"